refactor: Migrate Discord profile fields → Integrations tab
